Fix bug in distance method and add test
Add a unit test for the angularDistance method and add all
relevant libraries to the project.

Fix a bug in the angularDistance method (unit confusion between
radians and degrees). The same bug was present in the
computation of a random point withing a certain distance. This
is also fixed and a unit test has been added.